refactor: Nest provider-resources tools.
