+++ src/tools/index.ts
@@ -17,7 +17,6 @@ import "./financial-commitment-reports";
 import "./folders";
 import "./get-cost-provider-accounts";
 import "./get-financial-commitment-report";
-import "./get-provider-resource";
 import "./get-report-notification";
 import "./get-team";
 import "./get-teams";
@@ -25,9 +24,9 @@ import "./list-audit-logs";
 import "./list-cost-integrations";
 import "./list-cost-providers";
 import "./list-cost-services";
-import "./list-provider-resources";
 import "./list-report-notifications";
 import "./network-flow-reports";
+import "./provider-resources";
 import "./recommendation-views";
 import "./recommendations";
 import "./resource-reports";
