+++ src/tools/provider-resources/index.ts
@@ -1,0 +1,2 @@
+import "./get-provider-resource";
+import "./list-provider-resources";
