+++ src/tools/provider-resources/list-provider-resources.ts
@@ -1,7 +1,7 @@
 import z from "zod";
-import paginationData from "../utils/paginationData";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import paginationData from "../../utils/paginationData";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 List infrastructure provider resources (instances, volumes, load balancers, etc.) from your cloud accounts.
